Package structure overhaul
