+++ python/src/wallets/crossmint/goat_wallets/crossmint/__init__.py
@@ -1,12 +1,35 @@
 """Crossmint wallet implementation for GOAT SDK."""
 from .api_client import CrossmintWalletsAPI
-from .smart_wallet import SmartWalletClient, smart_wallet_factory
-from .custodial_solana_wallet import CustodialSolanaWalletClient, custodial_factory
+from .smart_wallet import smart_wallet_factory
+from .custodial_solana_wallet import custodial_factory
 
-__all__ = [
-    "CrossmintWalletsAPI",
-    "SmartWalletClient",
-    "smart_wallet_factory",
-    "CustodialSolanaWalletClient",
-    "custodial_factory"
-]
+"""CrossMint wallet implementation for GOAT SDK."""
+from typing import Dict, Any
+
+from .api_client import CrossmintWalletsAPI
+from .faucet_plugin import faucet_plugin
+from .mint_plugin import mint_plugin
+from .wallet_plugin import wallets_plugin
+from .custodial_solana_wallet import custodial_factory
+from .smart_wallet import smart_wallet_factory
+
+def crossmint(api_key: str) -> Dict[str, Any]:
+    """Initialize CrossMint SDK with API key.
+    
+    Args:
+        api_key: CrossMint API key
+    
+    Returns:
+        Dict containing CrossMint wallet and plugin factories
+    """
+    api_client = CrossmintWalletsAPI(api_key=api_key)
+
+    return {
+        "custodial": custodial_factory(api_client),
+        "smartwallet": smart_wallet_factory(api_client),
+        "faucet": faucet_plugin(api_client),
+        "mint": mint_plugin(api_client),
+        "wallets": wallets_plugin(api_client)
+    }
+
+__all__ = ["crossmint"]
