+++ python/src/wallets/crossmint/goat_wallets/crossmint/wallet_plugin.py
@@ -1,7 +1,7 @@
 from goat.classes import Chain
 from .base import PluginBase
 from .wallet import CrossmintWalletService
-from ...goat_wallets.crossmint.api_client import CrossmintWalletsAPI
+from .api_client import CrossmintWalletsAPI
 
 class WalletPlugin(PluginBase):
     def __init__(self, client: CrossmintWalletsAPI):
