+++ python/src/wallets/crossmint/pyproject.toml
@@ -9,8 +9,6 @@ homepage = "https://ohmygoat.dev/"
 repository = "https://github.com/goat-sdk/goat"
 packages = [
     { include = "goat_wallets/crossmint" },
-    { include = "goat_plugins/crossmint" },
-    { include = "crossmint" }
 ]
 
 [tool.poetry.dependencies]
